+++ CHANGES.md
@@ -7,6 +7,8 @@ Revision History
 - Added Git plugin support via: `git deps`
 - Removed '--no-clean' option (now the default) on 'install' and 'update'.
 - Added '--clean' option to delete ignored files on 'install' and 'update'.
+- Switched to 'install' rather than 'update' of nested dependencies.
+- Added '--all' option on 'update' to update all nested dependencies.
 
 0.4.1 (2015/09/24)
 ------------------